acl_yolov5_pt/onnx_exporter/export.py
"""
Yolov5 ONNX Exporter
Copyright 2021 Huawei Technologies Co., Ltd

Usage:
  $ export PYTHONPATH="$PWD" && python export.py --weights ./weights/yolov5s.pt --img 640 --batch 1 --simplify

CREATED:  2020-6-04 20:12:13
MODIFIED: 2021-11-01 01:48:45
"""

# -*- coding:utf-8 -*-
import argparse
import torch


def file_size(path):
    # Return file/dir size (MB)
    path = Path(path)
    if path.is_file():
        return path.stat().st_size / 1E6
    elif path.is_dir():
        return sum(f.stat().st_size for f in path.glob('**/*') if f.is_file()) / 1E6
    else:
        return 0.0


if __name__ == '__main__':
    parser = argparse.ArgumentParser()
    parser.add_argument('--weights', type=str, default='./yolov5s.pt', help='weights path')
    parser.add_argument('--img-size', nargs='+', type=int, default=[640, 640], help='image size')
    parser.add_argument('--batch-size', type=int, default=1, help='batch size')
    parser.add_argument('--simplify', action='store_true', help='ONNX: simplify model')
    opt = parser.parse_args()
    opt.img_size *= 2 if len(opt.img_size) == 1 else 1  # expand
    print(opt)

    # Input
    img = torch.zeros((opt.batch_size, 3, *opt.img_size))  # image size(1,3,320,192) iDetection

    # Load PyTorch model
    model = torch.load(opt.weights, map_location=torch.device('cpu'))['model'].float()
    model.eval()
    model.model[-1].export = True  # set Detect() layer export=True
    y = model(img)  # dry run

    # ONNX export
    try:
        import onnx

        print('\nStarting ONNX export with onnx %s...' % onnx.__version__)
        f = opt.weights.replace('.pt', '.onnx')  # filename
        model.fuse()  # only for ONNX
        torch.onnx.export(model, img, f, verbose=False, opset_version=11, input_names=['images'],
                          output_names=['classes', 'boxes'] if y is None else ['output'])

        # Checks
        onnx_model = onnx.load(f)  # load onnx model
        onnx.checker.check_model(onnx_model)  # check onnx model
        print(onnx.helper.printable_graph(onnx_model.graph))  # print a human readable model
        print('ONNX export success, saved as %s' % f)

         # Simplify
        if opt.simplify:
            try:
                import onnxsim

                print(f'ONNX: simplifying with onnx-simplifier {onnxsim.__version__}...')
                onnx_model, check = onnxsim.simplify(
                    onnx_model)
                assert check, 'assert check failed'
                onnx.save(onnx_model, f)
            except Exception as e:
                print(f'ONNX: simplifier failure: {e}')
        print(f'ONNX: export success, saved as {f} ({file_size(f):.1f} MB)')
        print(f"ONNX: run --dynamic ONNX model inference with: 'python detect.py --weights {f}'")
    except Exception as e:
        print('ONNX export failure: %s' % e)

    # Finish
    print('\nExport complete. Visualize with https://github.com/lutzroeder/netron.')
