acl_yolov5_pt/onnx_exporter/requirements.txt
# pip install -U -r requirements.txt
Cython
opencv-python-headless
matplotlib
pillow
tensorboard
PyYAML>=5.3
scipy
tqdm
onnx
onnx-simplifier
